Backend/requirements.txt
# Core dependencies
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.7.0

# MCP Protocol
mcp>=0.9.0

# LangGraph and LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0

# LLM providers (optional)
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0

# RAG System Dependencies
beautifulsoup4>=4.12.0
langchain-community>=0.3.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0

# Web Framework
flask
flask-cors>=3.0.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0
black>=24.0.0
ruff>=0.4.0

# Monitoring and Tracing
langsmith>=0.1.0